requirements.txt
flask
flask-cors
requests
flask-sqlalchemy
gunicorn
Pillow
